don't call $sel->start, it breaks things
t/01live.t fails with an error message telling that it couldn't start the Selenium
server for an unknown reason, despite the server running and having started a
browser. This server is also never stopped so the test never finishes because
it's waiting for child processes to exit.

It turns out that the $sel->start; call is not needed at all, and everything works
just as well without it. It's also not mentioned in Test::WWW::Selenium
documentation and examples.

The attached patch fixes this issue and gives a shot at stopping the Selenium
server by issuing a shutDown command before killing the Alien::SeleniumRC
child.

Tests run successfully for me and the module installs with this patch.